make initial loading faster
this was done by refactoring the frontends runtime settings to not be a
JSON file which had to be loaded from code but instead be javascript
code itself. By doing so, the browser can optimize loading times more
efficiently. This also removes the need for us to fetch the settings
twice (once before application init and once after).